Fix System.IO.FileSystem's PathHelpers for Unix
The FileSystem's PathHelpers internal class contains some utility functions for processing paths.  Some of these functions are copies of non-exposed Path members, while others are unique.  A bunch of those members have Windows-specifics coded into them, which is causing problems when dealing with various special paths on Unix, for example paths that end with double dots, relative paths that begin with a space, etc.

This commit splits PathHelpers into a PathHelpers.cs, PathHelpers.Windows.cs, and PathHelpers.Unix.cs, and implements the Unix functions appropriately.

It also fixes up a variety of tests that are now failing because they were expecting the Windows behavior.

+++ src/System.IO.FileSystem/tests/File/Create_str_i.cs
@@ -192,35 +192,38 @@ public class File_Create_str_i
 
             strLoc = "loc_89tbh_1";
 
-            //see VSWhidbey bug 103341
-            filName = Path.Combine(TestInfo.CurrentDirectory, Path.GetRandomFileName());
-            if (File.Exists(filName))
-                File.Delete(filName);
-            fs2 = File.Create(String.Format(" {0}", filName), 100);
-            iCountTestcases++;
-            if (!File.Exists(filName))
+            if (Interop.IsWindows) // tests are expecting whitespace at beginning of filename will be trimmed off, which it doesn't on Unix
             {
-                iCountErrors++;
-                printerr("Error_t87gy_1! File not created, file==" + filName);
-            }
-            fs2.Dispose();
-            File.Delete(filName);
-
-            strLoc = "loc_89tbh_3";
-
-            //see VSWhidbey bug 103341
-            filName = Path.Combine(" " + TestInfo.CurrentDirectory, " " + Path.GetRandomFileName());
-            if (File.Exists(filName))
+                //see VSWhidbey bug 103341
+                filName = Path.Combine(TestInfo.CurrentDirectory, Path.GetRandomFileName());
+                if (File.Exists(filName))
+                    File.Delete(filName);
+                fs2 = File.Create(String.Format(" {0}", filName), 100);
+                iCountTestcases++;
+                if (!File.Exists(filName))
+                {
+                    iCountErrors++;
+                    printerr("Error_t87gy_1! File not created, file==" + filName);
+                }
+                fs2.Dispose();
                 File.Delete(filName);
-            fs2 = File.Create(filName, 100);
-            iCountTestcases++;
-            if (!File.Exists(filName))
-            {
-                iCountErrors++;
-                printerr("Error_t87gy_3! File not created, file==" + filName);
+
+                strLoc = "loc_89tbh_3";
+
+                //see VSWhidbey bug 103341
+                filName = Path.Combine(" " + TestInfo.CurrentDirectory, " " + Path.GetRandomFileName());
+                if (File.Exists(filName))
+                    File.Delete(filName);
+                fs2 = File.Create(filName, 100);
+                iCountTestcases++;
+                if (!File.Exists(filName))
+                {
+                    iCountErrors++;
+                    printerr("Error_t87gy_3! File not created, file==" + filName);
+                }
+                fs2.Dispose();
+                File.Delete(filName);
             }
-            fs2.Dispose();
-            File.Delete(filName);
 
             if (File.Exists(filName))
                 File.Delete(filName);
